refactor: #5 기능별 파일 정리
