+++ app/src/main/AndroidManifest.xml
@@ -13,18 +13,18 @@
         android:theme="@style/Theme.NOWSOPTAndroid"
         tools:targetApi="31">
         <activity
-            android:name=".LoginActivity"
+            android:name=".Login.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity"
+        <activity android:name=".Main.MainActivity"
             android:exported="false"/>
-        <activity android:name=".SignupActivity"
+        <activity android:name=".Signup.SignupActivity"
             android:exported="false"/>
-        <activity android:name=".HomeActivity"
+        <activity android:name=".Home.HomeActivity"
             android:exported="false"/>
     </application>
