+++ app/src/main/java/com/sopt/now/Friend/Friend.kt
@@ -1,8 +1,8 @@
-package com.sopt.now
+package com.sopt.now.Friend
 
 import androidx.annotation.DrawableRes
 
-data class User(
+data class Friend(
     @DrawableRes val profileImage :Int,
     val nickname : String,
     val mbti : String
