+++ app/src/main/java/com/sopt/now/Friend/FriendViewHolder.kt
@@ -1,4 +1,4 @@
-package com.sopt.now
+package com.sopt.now.Friend
 
 import androidx.recyclerview.widget.RecyclerView
 import com.sopt.now.databinding.ItemFriendBinding
