+++ app/src/main/java/com/sopt/now/Mypage/MypageFragment.kt
@@ -1,4 +1,4 @@
-package com.sopt.now
+package com.sopt.now.Mypage
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,9 +6,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.sopt.now.databinding.FragmentMypageBinding
-import com.sopt.now.LoginActivity.Companion.ID
-import com.sopt.now.LoginActivity.Companion.MBTI
-import com.sopt.now.LoginActivity.Companion.NICKNAME
+import com.sopt.now.Login.LoginActivity.Companion.ID
+import com.sopt.now.Login.LoginActivity.Companion.MBTI
+import com.sopt.now.Login.LoginActivity.Companion.NICKNAME
 
 class MypageFragment : Fragment() {
     private var _binding : FragmentMypageBinding? = null
