+++ app/src/main/java/com/sopt/now/Search/SearchFragment.kt
@@ -1,4 +1,4 @@
-package com.sopt.now
+package com.sopt.now.Search
 
 import android.os.Bundle
 import android.view.LayoutInflater
