+++ app/src/main/java/com/sopt/now/User/UserViewHolder.kt
@@ -1,4 +1,4 @@
-package com.sopt.now
+package com.sopt.now.User
 
 import androidx.recyclerview.widget.RecyclerView
 import com.sopt.now.databinding.ItemUserBinding
