+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:padding="70dp"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".Main.MainActivity">
 
 
     <TextView
